requirements.txt
python-dotenv~=0.21.0
pydantic~=1.10.2
starlette==0.20.4
SQLAlchemy-Utils==0.39.0
SQLAlchemy==2.0.0

Flask~=2.1.3
aioflask~=0.4.0
alembic~=1.9.2
psycopg==3.1.8
